Cambios refactorizados Administracion de cuenta y Cambio de contraseña
