--- app/src/main/java/com/twulfz/ecualert/fragments/CuentaFragment.java
@@ -1,40 +1,0 @@
-package com.twulfz.ecualert.fragments;
-
-import android.content.Intent;
-import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.Button;
-
-import com.twulfz.ecualert.R;
-
-
-public class CuentaFragment extends Fragment {
-    Button btnCuenta;
-
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        // Inflate the layout for this fragment
-        View view = inflater.inflate(R.layout.fragment_cuenta, container, false);
-
-        // Obtener referencia al botón
-        btnCuenta = view.findViewById(R.id.btn_ccuenta);
-
-        // Configurar el OnClickListener para el botón
-        btnCuenta.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                // Crear una intención para abrir la actividad deseada
-                Intent intent = new Intent(getActivity(), Menucon.class);
-                startActivity(intent);
-            }
-        });
-
-        return view; }
-}
