+++ app/src/main/java/com/twulfz/ecualert/fragments/UserFragment.java
@@ -1,0 +1,39 @@
+package com.twulfz.ecualert.fragments;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.fragment.app.Fragment;
+
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.Button;
+
+import com.twulfz.ecualert.ActivityEditarContrasena;
+import com.twulfz.ecualert.R;
+
+
+public class UserFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_user, container, false);
+    }
+
+    @Override
+    public void onViewCreated(View view, Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        Button btnEditPass = view.findViewById(R.id.btn_edit_pass);
+        btnEditPass.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+                Intent intent = new Intent(getContext(), ActivityEditarContrasena.class);
+                startActivity(intent);
+            }
+        });
+    }
+}
