+++ app/src/main/res/drawable/ic_report.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="800dp"
-    android:height="800dp"
+    android:width="32dp"
+    android:height="32dp"
     android:viewportWidth="392.53"
     android:viewportHeight="392.53">
   <path
