+++ app/src/main/res/drawable/ic_user.xml
@@ -1,5 +1,5 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="24dp" android:viewportHeight="24" android:viewportWidth="24" android:width="24dp">
       
-    <path android:fillColor="#FFF" android:pathData="M12,22C6.477,22 2,17.523 2,12S6.477,2 12,2s10,4.477 10,10 -4.477,10 -10,10zM7.013,18.256A7.966,7.966 0,0 0,12 20c1.97,0 3.773,-0.712 5.167,-1.892A6.979,6.979 0,0 0,12.16 16a6.981,6.981 0,0 0,-5.147 2.256zM5.616,16.82A8.975,8.975 0,0 1,12.16 14a8.972,8.972 0,0 1,6.362 2.634,8 8,0 1,0 -12.906,0.187zM12,13a4,4 0,1 1,0 -8,4 4,0 0,1 0,8zM12,11a2,2 0,1 0,0 -4,2 2,0 0,0 0,4z"/>
-    
+    <path android:fillColor="#000000"  android:pathData="M12,22C6.477,22 2,17.523 2,12S6.477,2 12,2s10,4.477 10,10 -4.477,10 -10,10zM7.013,18.256A7.966,7.966 0,0 0,12 20c1.97,0 3.773,-0.712 5.167,-1.892A6.979,6.979 0,0 0,12.16 16a6.981,6.981 0,0 0,-5.147 2.256zM5.616,16.82A8.975,8.975 0,0 1,12.16 14a8.972,8.972 0,0 1,6.362 2.634,8 8,0 1,0 -12.906,0.187zM12,13a4,4 0,1 1,0 -8,4 4,0 0,1 0,8zM12,11a2,2 0,1 0,0 -4,2 2,0 0,0 0,4z"/>
+
 </vector>
