+++ app/src/main/res/drawable/user.xml
@@ -1,0 +1,7 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android" android:height="512dp" android:viewportHeight="24" android:viewportWidth="24" android:width="512dp">
+      
+    <path android:fillColor="#000000" android:fillType="evenOdd" android:pathData="M8.25,9C8.25,6.929 9.929,5.25 12,5.25C14.071,5.25 15.75,6.929 15.75,9C15.75,11.071 14.071,12.75 12,12.75C9.929,12.75 8.25,11.071 8.25,9ZM12,6.75C10.757,6.75 9.75,7.757 9.75,9C9.75,10.243 10.757,11.25 12,11.25C13.243,11.25 14.25,10.243 14.25,9C14.25,7.757 13.243,6.75 12,6.75Z"/>
+      
+    <path android:fillColor="#000000" android:fillType="evenOdd" android:pathData="M1.25,12C1.25,6.063 6.063,1.25 12,1.25C17.937,1.25 22.75,6.063 22.75,12C22.75,17.937 17.937,22.75 12,22.75C6.063,22.75 1.25,17.937 1.25,12ZM12,2.75C6.891,2.75 2.75,6.891 2.75,12C2.75,14.546 3.778,16.851 5.442,18.524C5.622,17.55 5.977,16.633 6.688,15.895C7.753,14.792 9.454,14.25 12,14.25C14.546,14.25 16.247,14.792 17.312,15.895C18.023,16.633 18.377,17.551 18.558,18.524C20.222,16.851 21.25,14.546 21.25,12C21.25,6.891 17.109,2.75 12,2.75ZM17.194,19.655C17.092,18.444 16.829,17.555 16.232,16.936C15.582,16.263 14.379,15.75 12,15.75C9.621,15.75 8.418,16.263 7.768,16.936C7.171,17.555 6.908,18.443 6.806,19.655C8.287,20.662 10.075,21.25 12,21.25C13.925,21.25 15.713,20.662 17.194,19.655Z"/>
+    
+</vector>
