+++ app/src/main/res/layout/activity_editar_contrasena.xml
@@ -1,0 +1,137 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/main"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".ActivityEditarContrasena"
+    android:background="@color/white">
+
+
+    <ImageView
+        android:id="@+id/imageView"
+        android:layout_width="180dp"
+        android:layout_height="180dp"
+        android:background="@drawable/circle_white_border"
+        android:src="@drawable/logo"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintHorizontal_bias="0.497"
+        app:layout_constraintStart_toStartOf="parent"
+        android:layout_marginTop="45dp"
+        app:layout_constraintTop_toTopOf="parent" />
+
+    <androidx.cardview.widget.CardView
+        android:id="@+id/cardView"
+        app:cardElevation="-1dp"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginHorizontal="30dp"
+        android:layout_marginTop="25dp"
+        app:cardCornerRadius="20dp"
+        app:layout_constraintTop_toBottomOf="@+id/imageView">
+
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            android:orientation="vertical"
+            android:background="@drawable/border_black"
+            android:paddingTop="32dp"
+            android:paddingHorizontal="24dp">
+
+
+            <TextView
+                android:id="@+id/textView"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Ingrese la nueva contraseña"
+                android:textAlignment="center"
+                android:textColor="@color/black"
+                style="@style/EditText"
+
+                android:textSize="20dp"
+                android:textStyle="bold"
+                />
+
+            <EditText
+                android:id="@+id/Edit_contra_Actual"
+                android:layout_width="match_parent"
+                android:layout_height="50dp"
+                style="@style/EditText"
+                android:layout_marginTop="40dp"
+                android:background="@drawable/border_black"
+                android:drawablePadding="8dp"
+                android:inputType="textPassword"
+                android:textColorHint="@color/black"
+                android:hint="Contraseña actual"
+                android:padding="8dp"
+                android:textColor="@color/black"
+                />
+
+            <EditText
+                android:id="@+id/Edit_nue_Contra"
+                android:layout_width="match_parent"
+                android:layout_height="50dp"
+                android:layout_marginTop="40dp"
+                android:background="@drawable/border_black"
+                android:drawablePadding="8dp"
+                style="@style/EditText"
+                android:inputType="textPassword"
+                android:textColorHint="@color/black"
+                android:hint="Nueva contraseña"
+                android:padding="8dp"
+                android:textColor="@color/black"
+                android:textColorHighlight="@color/cardview_dark_background" />
+            <EditText
+                android:id="@+id/Edit_conf_Contra"
+                android:layout_width="match_parent"
+                android:layout_height="50dp"
+                android:layout_marginTop="40dp"
+                android:background="@drawable/border_black"
+                android:drawablePadding="8dp"
+                android:hint="Confirmar contraseña"
+                android:inputType="textPassword"
+                android:textColorHint="@color/black"
+                android:padding="8dp"
+                style="@style/EditText"
+                android:textColor="@color/black"
+                android:textColorHighlight="@color/cardview_dark_background" />
+
+            <LinearLayout
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:gravity="center"
+                android:orientation="horizontal"
+                android:paddingVertical="12dp">
+
+                <Button
+                    android:id="@+id/btn_edit"
+                    android:layout_width="150dp"
+                    android:layout_height="wrap_content"
+                    android:layout_marginEnd="6dp"
+                    android:layout_marginBottom="32dp"
+                    android:backgroundTint="@color/secondary_color"
+                    android:text="Modificar"
+                    android:padding="16dp"
+                    android:layout_marginTop="48dp"
+                    android:layout_weight="1"
+                    android:gravity="center"
+                    style="@style/EditText"
+                    android:textStyle="bold"
+                    android:textColor="@color/font_color"
+                    android:textSize="18sp"
+                    app:cornerRadius="20dp"
+
+                    />
+
+
+            </LinearLayout>
+
+
+        </LinearLayout>
+    </androidx.cardview.widget.CardView>
+
+
+</androidx.constraintlayout.widget.ConstraintLayout>
