+++ app/src/main/res/values/strings.xml
@@ -2,14 +2,8 @@
     <string name="app_name">Ecualert</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
-<<<<<<< HEAD
     <string name="title_map">Mapa</string>
     <string name="title_home">Inicio</string>
     <string name="title_user">Usuario</string>
-=======
-    <string name="title_activity_botton_nave_menu_admcuenta">botton_nave_menu_admcuenta</string>
-    <string name="title_home">Home</string>
-    <string name="title_dashboard">Dashboard</string>
-    <string name="title_notifications">Notifications</string>
->>>>>>> 501a6af42a2b9b9ee76bf54313eacee8f253d7ad
+
 </resources>
