+++ app/src/main/res/values/styles.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <style name="EditText">
+        <item name="android:fontFamily">sans-serif</item>
+    </style>
+</resources>
